(process) Lift ranking calculations into an external process ... to not run out of memory
