+++ code/index/java/nu/marginalia/index/searchset/construction/ConnectivitySetsCalculator.java
@@ -1,0 +1,155 @@
+package nu.marginalia.index.searchset.construction;
+
+import com.google.inject.Inject;
+import com.zaxxer.hikari.HikariDataSource;
+import it.unimi.dsi.fastutil.ints.Int2ByteOpenHashMap;
+import it.unimi.dsi.fastutil.ints.Int2IntOpenHashMap;
+import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
+import it.unimi.dsi.fastutil.ints.IntArrayList;
+import it.unimi.dsi.fastutil.ints.IntList;
+import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
+import it.unimi.dsi.fastutil.ints.IntSet;
+import nu.marginalia.IndexLocations;
+import nu.marginalia.api.linkgraph.AggregateLinkGraphClient;
+import nu.marginalia.db.DomainRankingSetsService;
+import nu.marginalia.index.searchset.RankingSearchSet;
+import nu.marginalia.index.searchset.connectivity.ConnectivityMapFile;
+import nu.marginalia.index.searchset.connectivity.ConnectivityView;
+import nu.marginalia.index.searchset.connectivity.DomainSetConnectivity;
+import nu.marginalia.storage.FileStorageService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
+import java.nio.file.Path;
+import java.sql.SQLException;
+
+/** Calculates the connectivity map, which classifies each indexed domain by how it connects
+ * to the domains in the origin ranking set.  The result is persisted to the search sets
+ * directory, where the index service picks it up.
+ */
+public class ConnectivitySetsCalculator {
+
+    private static final Logger logger = LoggerFactory.getLogger(ConnectivitySetsCalculator.class);
+
+    private final AggregateLinkGraphClient graphClient;
+    private final HikariDataSource dataSource;
+    private final String sourceSetName;
+    private final Path searchSetsBase;
+    private final Path filePath;
+
+    @Inject
+    public ConnectivitySetsCalculator(AggregateLinkGraphClient graphClient,
+                                      FileStorageService fileStorageService,
+                                      HikariDataSource dataSource
+                                      ) {
+        this.graphClient = graphClient;
+        this.dataSource = dataSource;
+        this.sourceSetName = "SMALL";
+        this.searchSetsBase = IndexLocations.getSearchSetsPath(fileStorageService);
+
+        filePath = DomainRankingSetsService.setFileName(searchSetsBase, ConnectivityMapFile.FILE_NAME);
+    }
+
+    public ConnectivityView recalculate() {
+        Int2ObjectOpenHashMap<IntList> outgoingLinks = new Int2ObjectOpenHashMap<>();
+        Int2ObjectOpenHashMap<IntList> inboudLinks = new Int2ObjectOpenHashMap<>();
+        Int2IntOpenHashMap linkCount = new Int2IntOpenHashMap();
+
+        IntOpenHashSet indexedDomains = new IntOpenHashSet();
+
+        try {
+            // Load all relevant IDs so we can exclude links to non-indexed domains
+
+            try (var conn = dataSource.getConnection();
+                 var stmt = conn.prepareStatement("SELECT ID FROM EC_DOMAIN WHERE NODE_AFFINITY>0")) {
+                stmt.setFetchSize(10_000);
+                var rs = stmt.executeQuery();
+                while (rs.next()) {
+                    indexedDomains.add(rs.getInt(1));
+                }
+            }
+
+            // Load the link graph
+            var iter = graphClient.getAllDomainLinks().iterator();
+            while (iter.advance()) {
+                int src = iter.source();
+                int dest = iter.dest();
+
+                if (indexedDomains.contains(src) && indexedDomains.contains(dest)) {
+                    inboudLinks.computeIfAbsent(src, v -> new IntArrayList()).add(dest);
+                    outgoingLinks.computeIfAbsent(dest, v -> new IntArrayList()).add(src);
+                }
+            }
+
+            var originSet = new RankingSearchSet(sourceSetName,
+                    DomainRankingSetsService.setFileName(searchSetsBase, sourceSetName));
+
+            var idsList = originSet.domainIds();
+
+            IntSet direct = new IntOpenHashSet(idsList);
+            IntSet reachable = new IntOpenHashSet();
+            IntSet reachable2 = new IntOpenHashSet();
+            IntSet linking = new IntOpenHashSet();
+            IntSet linking2 = new IntOpenHashSet();
+            IntSet bidi = new IntOpenHashSet();
+            IntSet bidi2 = new IntOpenHashSet();
+
+            for (int id : idsList) {
+                var outgoing = outgoingLinks.getOrDefault(id, IntList.of());
+                for (int outId: outgoing) linkCount.addTo(outId, 1);
+
+                var inbound = inboudLinks.getOrDefault(id, IntList.of());
+                for (int inId: inbound) linkCount.addTo(inId, 1);
+
+                reachable.addAll(inbound);
+                linking.addAll(outgoing);
+            }
+
+            reachable.removeAll(direct);
+            linking.removeAll(direct);
+
+            bidi.addAll(reachable);
+            bidi.retainAll(linking);
+
+            reachable2.addAll(reachable);
+            linking2.addAll(linking);
+            bidi2.addAll(bidi);
+
+            reachable2.removeIf(id -> linkCount.get(id) < 5);
+            linking2.removeIf(id -> linkCount.get(id) < 5);
+            bidi2.removeIf(id -> linkCount.get(id) < 5);
+
+            Int2ByteOpenHashMap newConnectivity = new Int2ByteOpenHashMap();
+
+            direct.forEach(id -> newConnectivity.put(id, (byte) DomainSetConnectivity.DIRECT.ordinal()));
+
+            bidi2.forEach(id -> newConnectivity.putIfAbsent(id, (byte) DomainSetConnectivity.BIDI_HOT.ordinal()));
+            reachable2.forEach(id -> newConnectivity.putIfAbsent(id, (byte) DomainSetConnectivity.REACHABLE_HOT.ordinal()));
+            linking2.forEach(id -> newConnectivity.putIfAbsent(id, (byte) DomainSetConnectivity.LINKING_HOT.ordinal()));
+
+            bidi.forEach(id -> newConnectivity.putIfAbsent(id, (byte) DomainSetConnectivity.BIDI.ordinal()));
+            reachable.forEach(id -> newConnectivity.putIfAbsent(id, (byte) DomainSetConnectivity.REACHABLE.ordinal()));
+            linking.forEach(id -> newConnectivity.putIfAbsent(id, (byte) DomainSetConnectivity.LINKING.ordinal()));
+
+            ConnectivityMapFile.write(filePath, newConnectivity);
+
+            return new ConnectivityView(newConnectivity);
+        }
+        catch (SQLException | IOException ex) {
+            logger.error("Failed to calculate connectivity sets", ex);
+
+            return loadPrevious();
+        }
+    }
+
+    private ConnectivityView loadPrevious() {
+        try {
+            return new ConnectivityView(ConnectivityMapFile.read(filePath));
+        }
+        catch (IOException ex) {
+            return new ConnectivityView(new Int2ByteOpenHashMap());
+        }
+    }
+
+}
