+++ code/index/test/nu/marginalia/index/IndexQueryServiceIntegrationTestModule.java
@@ -4,7 +4,6 @@ import com.google.inject.AbstractModule;
 import nu.marginalia.IndexLocations;
 import nu.marginalia.index.journal.IndexJournal;
 import nu.marginalia.index.journal.IndexJournalSlopWriter;
-import nu.marginalia.index.searchset.DomainRankings;
 import nu.marginalia.index.searchset.SearchSetAny;
 import nu.marginalia.index.searchset.SearchSetsService;
 import nu.marginalia.linkdb.docs.DocumentDbReader;
@@ -80,7 +79,6 @@ public class IndexQueryServiceIntegrationTestModule extends AbstractModule {
             SearchSetsService setsServiceMock = Mockito.mock(SearchSetsService.class);
             when(setsServiceMock.getSearchSetByName("")).thenReturn(new SearchSetAny());
             when(setsServiceMock.getSearchSetByName("NONE")).thenReturn(new SearchSetAny());
-            when(setsServiceMock.getDomainRankings()).thenReturn(new DomainRankings());
             bind(SearchSetsService.class).toInstance(setsServiceMock);
 
             bind(ServiceEventLog.class).toInstance(Mockito.mock(ServiceEventLog.class));
