+++ code/processes/process-mq-api/java/nu/marginalia/mqapi/ranking/CreateRankingsRequest.java
@@ -1,0 +1,5 @@
+package nu.marginalia.mqapi.ranking;
+
+public record CreateRankingsRequest(RankingsName rankingsName)
+{
+}
