+++ code/processes/process-mq-api/java/nu/marginalia/mqapi/ranking/RankingsName.java
@@ -1,0 +1,6 @@
+package nu.marginalia.mqapi.ranking;
+
+public enum RankingsName {
+    PRIMARY,
+    SECONDARY
+}
